Many Updates to features and bugs
Bug #2 resolved in this issue.

New difficulty level 4 added, with more application questions.

Ability to import your own music into the game for both during the game
and at the menu.

Readme file added to inform viewers of what they are seeing.

Began work on multiplayer function of the game.